Fix elapsed time computation on windows
Assumed CLOCKS_PER_SEC was 1000000, but it's 1000 on windows.
So search time on windows was wrong by a factor of 1000.

+++ source/Perft.cpp
@@ -48,7 +48,7 @@ void perft_benchmark(string fen, int depth)
     long move_count = perft(board, depth);
     clock_t toc = clock();
 
-    long double elapsed_secs = static_cast<long double>(toc - tic) / CLOCKS_PER_SEC;
+    int elapsed_secs = int(double(toc - tic) / CLOCKS_PER_SEC);
     cout << "time: " << elapsed_secs << "s" << endl;
     cout << "moves: " << move_count << endl;
     cout << "moves per second: " << int(move_count / elapsed_secs) << endl;
